Rename repo_query to repo_manage
