+++ pkgs/repo_manage/README.md
@@ -1,0 +1,20 @@
+Various and miscellaneous commands to query dart-lang github repos.
+
+## Usage
+
+```
+Usage: report <command> [arguments]
+
+Global options:
+-h, --help    Print this usage information.
+
+Available commands:
+  branches          Show the default branch names of Dart and Flutter repos.
+  labels            Report on the various labels in use by dart-lang repos.
+  labels-update     Audit and update the labels used by dart-lang repos.
+  links             List useful GitHub query urls.
+  transfer-issues   Bulk transfer issues from one repo to another.
+  weekly            Run a week-based report on repo status and activity.
+
+Run "report help <command>" for more information about a command.
+```
