+++ pkgs/repo_manage/analysis_options.yaml
