+++ pkgs/repo_manage/lib/weekly.dart
